# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "API Documentation",
    "description": "RESTful API Documentation - User, Role, Category Management System",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/api",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "JWT Authorization header using the Bearer scheme. Example: \"Bearer {token}\""
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/roles/": {
      "get": {
        "tags": [
          "Roles"
        ],
        "summary": "Get all roles",
        "description": "Retrieve a list of all roles in the system",
        "parameters": [
          {
            "name": "accept-language",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Roles retrieved successfully"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/roles/permissions": {
      "get": {
        "tags": [
          "Roles"
        ],
        "summary": "Get all available permissions",
        "description": "Retrieve all system permissions for role assignment",
        "parameters": [
          {
            "name": "accept-language",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Permissions retrieved successfully"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/roles/role_privileges": {
      "get": {
        "tags": [
          "Roles"
        ],
        "summary": "Get privileges for a specific role",
        "description": "Retrieve all privileges assigned to a role",
        "parameters": [
          {
            "name": "accept-language",
            "in": "header",
            "type": "string"
          },
          {
            "name": "role_id",
            "in": "query",
            "description": "Role ID",
            "required": true,
            "type": "string",
            "example": "507f1f77bcf86cd799439012"
          }
        ],
        "responses": {
          "200": {
            "description": "Role privileges retrieved successfully"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/roles/add": {
      "post": {
        "tags": [
          "Roles"
        ],
        "summary": "Add a new role",
        "description": "Create a new role with permissions",
        "parameters": [
          {
            "name": "accept-language",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Role information",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "role_name": {
                  "type": "string",
                  "example": "Admin"
                },
                "is_active": {
                  "type": "boolean",
                  "example": true
                },
                "permissions": {
                  "type": "array",
                  "example": [
                    "user_view",
                    "user_add",
                    "user_update",
                    "user_delete"
                  ],
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Role created successfully"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/roles/update": {
      "post": {
        "tags": [
          "Roles"
        ],
        "summary": "Update an existing role",
        "description": "Update role information and permissions",
        "parameters": [
          {
            "name": "accept-language",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Role update information",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "string",
                  "example": "507f1f77bcf86cd799439012"
                },
                "role_name": {
                  "type": "string",
                  "example": "Super Admin"
                },
                "is_active": {
                  "type": "boolean",
                  "example": true
                },
                "permissions": {
                  "type": "array",
                  "example": [
                    "user_view",
                    "user_add",
                    "user_update",
                    "user_delete"
                  ],
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Role updated successfully"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/roles/delete": {
      "post": {
        "tags": [
          "Roles"
        ],
        "summary": "Delete a role",
        "description": "Delete a role and its associated privileges",
        "parameters": [
          {
            "name": "accept-language",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Role ID to delete",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "string",
                  "example": "507f1f77bcf86cd799439012"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Role deleted successfully"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/routes/": {
      "post": {
        "tags": [
          "Upload"
        ],
        "summary": "Upload a file",
        "description": "Upload a single file to the server",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "accept-language",
            "in": "header",
            "type": "string"
          },
          {
            "name": "file",
            "in": "formData",
            "type": "file",
            "required": true,
            "description": "File to upload"
          }
        ],
        "responses": {
          "200": {
            "description": "File uploaded successfully"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
...
}